Reorganized torben files for new structure
