+++ torben/server/dijkstra.js
