+++ torben/server/weightMap.js
